requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
_tflow_select=2.3.0=mkl
absl-py=0.8.1=pypi_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.0=pypi_0
astunparse=1.6.3=pypi_0
async-generator=1.10=pypi_0
attrs=19.3.0=py_0
backcall=0.2.0=pypi_0
blas=1.1=openblas
bleach=3.2.0=pypi_0
bravado=10.6.2=pypi_0
bravado-core=5.17.0=pypi_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.6.20=hecda079_0
cachetools=3.1.1=pypi_0
cairo=1.14.12=h8948797_3
certifi=2020.6.20=py36h9f0ad1d_0
cffi=1.12.3=py36h2e261b9_0
click=7.1.2=pypi_0
cudatoolkit=9.2=0
cxxfilt=0.2.1=py36h831f99a_1
cycler=0.10.0=py_1
defcon=0.7.2=pypi_0
defusedxml=0.6.0=pypi_0
densratio=0.2.2=pypi_0
dill=0.3.2=pypi_0
dominate=2.5.1=pypi_0
easydict=1.9=pypi_0
easydl=2.0.7=pypi_0
efficientnet-pytorch=0.7.0=pypi_0
entrypoints=0.3=pypi_0
faiss=1.5.3=pypi_0
ffmpeg=4.0=hcdf2ecd_0
fontconfig=2.13.1=he4413a7_1000
fontmath=0.6.0=pypi_0
fonttools=4.12.1=pypi_0
freeglut=3.0.0=hf484d3e_5
freetype=2.9.1=h8a8886c_1
future=0.18.2=pypi_0
gast=0.2.2=pypi_0
gitdb=4.0.5=pypi_0
gitpython=3.1.8=pypi_0
glib=2.63.1=h5a9c865_0
glyphslib=2.4.0=pypi_0
google-auth=1.7.1=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
google-pasta=0.1.8=pypi_0
gputil=1.4.0=pypi_0
graphite2=1.3.13=h23475e2_0
grpcio=1.25.0=pypi_0
h5py=2.8.0=py36h989c5e5_3
harfbuzz=1.8.8=hffaf4a1_0
hdbscan=0.8.22=py36hd352d35_1
hdf5=1.10.2=hba1933b_1
icu=58.2=hf484d3e_1000
importlib-metadata=1.6.0=py36h9f0ad1d_0
importlib_metadata=1.6.0=0
intel-openmp=2019.4=243
ipdb=0.13.3=pypi_0
ipykernel=5.3.4=pypi_0
ipython=7.16.1=pypi_0
ipython-genutils=0.2.0=pypi_0
jasper=2.0.14=h07fcdf6_1
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
joblib=0.13.2=py36_0
jpeg=9b=h024ee3a_2
jsonpatch=1.26=pypi_0
jsonpointer=2.0=pypi_0
jsonref=0.2=pypi_0
jsonschema=3.2.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-core=4.6.3=pypi_0
jupyterlab-pygments=0.1.1=pypi_0
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py36hc9558a2_0
kmeans-pytorch=0.3=pypi_0
libblas=3.8.0=11_openblas
libcblas=3.8.0=11_openblas
libedit=3.1.20181209=hc058e9b_0
libfaiss=1.6.3=h49bdc20_1_cuda
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.3.0=h24d8f2e_16
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libgomp=9.3.0=h24d8f2e_16
libiconv=1.15=h516909a_1005
liblapack=3.8.0=11_openblas
libopenblas=0.3.6=h5a2b251_2
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.4=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.0.10=h2733197_2
libuuid=2.32.1=h14c3975_1000
libvpx=1.7.0=h439df22_0
libxcb=1.13=h1bed415_1
libxml2=2.9.9=h13577e0_2
markdown=3.1.1=py36_0
markupsafe=1.1.1=pypi_0
matplotlib=3.1.0=py36_0
matplotlib-base=3.1.0=py36h5f35d83_0
mistune=0.8.4=pypi_0
mkl=2019.4=243
mkl-service=2.0.2=py36h7b6447c_0
mlconfig=0.1.0=pypi_0
mock=4.0.2=py_0
monotonic=1.5=pypi_0
more-itertools=8.2.0=py_0
msgpack=0.6.2=pypi_0
msgpack-numpy=0.4.4.3=pypi_0
msgpack-python=0.5.6=pypi_0
munch=2.5.0=py_0
mutatormath=3.0.1=pypi_0
nbclient=0.5.0=pypi_0
nbconvert=6.0.3=pypi_0
nbformat=5.0.7=pypi_0
ncurses=6.1=he6710b0_1
neptune-client=0.4.120=pypi_0
neptune-notebooks=0.0.16=pypi_0
nest-asyncio=1.4.0=pypi_0
ninja=1.9.0=py36hfd86e86_0
notebook=6.1.4=pypi_0
numpy=1.18.1=py36h94c655d_0
numpy-base=1.18.1=py36h2f8d375_1
nvidia-apex=0.1=py36h88da601_1
oauthlib=3.1.0=pypi_0
olefile=0.46=py36_0
openblas=0.3.3=h9ac9557_1001
opencv=3.4.2=py36h6fd60c2_1
openssl=1.1.1g=h516909a_1
opt-einsum=3.1.0=pypi_0
packaging=20.1=py_0
pandas=0.25.1=py36he6710b0_0
pandocfilters=1.4.2=pypi_0
parso=0.7.1=pypi_0
path=15.0.0=pypi_0
path-py=12.5.0=pypi_0
pathlib2=2.3.5=pypi_0
pcre=8.43=he6710b0_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=6.1.0=py36h34e0f95_0
pip=19.2.2=py36_0
pixman=0.38.0=h7b6447c_0
pluggy=0.12.0=py_0
pretrainedmodels=0.7.4=pypi_0
progress=1.5=py36_0
prometheus-client=0.8.0=pypi_0
prompt-toolkit=3.0.5=pypi_0
protobuf=3.9.2=pypi_0
psutil=5.6.5=pypi_0
ptyprocess=0.6.0=pypi_0
py=1.8.1=py_0
py-opencv=3.4.2=py36hb342d67_1
py3nvml=0.2.6=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.7=pypi_0
pycparser=2.19=py36_0
pygments=2.6.1=pypi_0
pyjwt=1.7.1=pypi_0
pykeops=1.3=pypi_0
pyparsing=2.4.2=py_0
pyqt=4.11.4=py36_3
pyrsistent=0.17.3=pypi_0
pytest=5.4.1=py36h9f0ad1d_0
python=3.6.9=h265db76_0
python-dateutil=2.8.0=py_0
python_abi=3.6=1_cp36m
pytorch=1.2.0=py3.6_cuda9.2.148_cudnn7.6.2_0
pytz=2019.2=py_0
pywsl=0.1.4=pypi_0
pyyaml=5.1.2=py36h7b6447c_0
pyzmq=18.1.1=pypi_0
qt=4.8.7=2
readline=7.0=h7b6447c_5
requests-oauthlib=1.3.0=pypi_0
rfc3987=1.3.8=pypi_0
rsa=4.0=pypi_0
scikit-learn=0.20.4=py36_blas_openblashebff5e3_0
scipy=1.4.1=pypi_0
seaborn=0.11.0=pypi_0
send2trash=1.5.0=pypi_0
setuptools=41.0.1=py36_0
simplejson=3.17.2=pypi_0
sip=4.18=py36_1
six=1.12.0=py36_0
sklearn=0.0=pypi_0
smmap=3.0.4=pypi_0
sqlite=3.29.0=h7b6447c_0
strict-rfc3339=0.7=pypi_0
swagger-spec-validator=2.7.3=pypi_0
tabulate=0.8.5=pypi_0
tensorboard=2.2.2=pypi_0
tensorboard-plugin-wit=1.6.0.post3=pypi_0
tensorboardx=1.8=pypi_0
tensorflow=2.0.0=pypi_0
tensorflow-base=1.13.1=mkl_py36h7ce6ba3_0
tensorflow-estimator=2.2.0=pypi_0
tensorflow-gpu=2.2.0=pypi_0
tensorpack=0.9.8=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.8.3=pypi_0
testpath=0.4.4=pypi_0
tk=8.6.9=hed695b0_1002
torchfile=0.1.0=pypi_0
torchvision=0.4.0=py36_cu92
tornado=6.0.3=py36h516909a_0
tqdm=4.32.1=py_0
traitlets=4.3.3=pypi_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.24.3=pypi_0
visdom=0.1.8.9=pypi_0
wcwidth=0.1.9=pyh9f0ad1d_0
webcolors=1.11.1=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=0.16.0=pypi_0
wheel=0.33.4=py36_0
xmltodict=0.12.0=pypi_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
zipp=3.1.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
